moved `move` method from utils.py to static_funcs.py, added `download_external_files`
